+++ src/components/dice/Dice.jsx
@@ -6,8 +6,8 @@ import "./dice.scss";
 import NavbarDice from "./NavbarDice";
 
 const Dice = () => {
-  const {restaurantes}=useSelector((store)=>store.restaurantStore)
-  const disptach=useDispatch()
+  const { restaurantes } = useSelector((store) => store.restaurantStore);
+  const disptach = useDispatch();
   // const lista = [
   //   "pizzas juan",
   //   "pizza picolo",
@@ -22,18 +22,14 @@ const Dice = () => {
   const [dado, setDado] = useState(false);
   useEffect(() => {
     if (!restaurantes.length) {
-      disptach(actionGetrestaurantesAsync())
-   console.log(restaurantes);
+      disptach(actionGetrestaurantesAsync());
+      console.log(restaurantes);
     }
-   
-   
-  }, [restaurantes])
+  }, [restaurantes]);
   console.log(restaurantes);
-  const lista= restaurantes.map((item,index)=>(
-    item.name
-  ))
-  
-  const changeDado=()=>{
+  const lista = restaurantes.map((item, index) => item.name);
+
+  const changeDado = () => {
     setDado(true);
     setTimeout(() => {
       const x = Math.floor(Math.random() * lista.length);
@@ -46,33 +42,35 @@ const Dice = () => {
     setTimeout(() => {
       setDado(false);
     }, 3500);
-  
-  }
+  };
   return (
     <>
       <NavbarDice />
-      {lista.length?<div className="contenedor">
-        {dado ? (
-          <div className={"dado"}>
-            <div className="lado uno"></div>
-            <div className="lado dos"></div>
-            <div className="lado tres"></div>
-            <div className="lado cuatro"></div>
-            <div className="lado cinco"></div>
-            <div className="lado seis"></div>
-          </div>
-        ) : (
-          <div onClick={changeDado} className={"dados"}>
-            <div className="lado uno"></div>
-            <div className="lado dos"></div>
-            <div className="lado tres"></div>
-            <div className="lado cuatro"></div>
-            <div className="lado cinco"></div>
-            <div className="lado seis"></div>
-          </div>
-        )}
-      </div>:<h1> LOADING ...</h1>}
-      
+      {lista.length ? (
+        <div className="contenedor">
+          {dado ? (
+            <div className={"dado"}>
+              <div className="lado uno"></div>
+              <div className="lado dos"></div>
+              <div className="lado tres"></div>
+              <div className="lado cuatro"></div>
+              <div className="lado cinco"></div>
+              <div className="lado seis"></div>
+            </div>
+          ) : (
+            <div onClick={changeDado} className={"dados"}>
+              <div className="lado uno"></div>
+              <div className="lado dos"></div>
+              <div className="lado tres"></div>
+              <div className="lado cuatro"></div>
+              <div className="lado cinco"></div>
+              <div className="lado seis"></div>
+            </div>
+          )}
+        </div>
+      ) : (
+        <h1> LOADING ...</h1>
+      )}
     </>
   );
 };
