+++ src/components/dice/dice.scss
@@ -24,8 +24,16 @@ $color-sesion: #f9943b;
     @include buttons(1);
     background-color: $color-sesion;
   }
+  &__figure {
+    width: 100%;
+    height: 40px;
+  }
+  &__imgLogo {
+    width: 50px;
+  }
   &__img {
-    width: 30px;
+    height: 30px;
+    margin-right: 30px;
   }
   &__section {
     display: flex;
