+++ src/components/home/Home.jsx
@@ -1,86 +1,15 @@
-import React, { useEffect, useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { Link, useNavigate } from "react-router-dom";
-import Swal from "sweetalert2";
+import React from "react";
 import ControlledCarousel from "../carousel/Carousel";
 import Footer from "../footer/Footer";
-import { actionGetrestaurantesAsync } from "../../redux/actions/restaurantsActions";
-import {
-  actionLogout,
-  actionUserLogOutAsync,
-} from "../../redux/actions/userActions";
 import NavbarDice from "../dice/NavbarDice";
 import "./home.scss";
+import { useNavigate } from "react-router-dom";
 
-const Home = ({ isAuthentication }) => {
-  const lista = [
-    "pizzas juan",
-    "pizza picolo",
-    "pizza carlota",
-    "carlos pizza",
-    "pizza picolo",
-    "pizza dominos",
-    "pizza daikmaku",
-    "pizza lauras",
-    "kakos pizza",
-  ];
-  const [dado, setDado] = useState(false);
-  const [userFunctions, setUserFunctions] = useState(false);
-  const disptach = useDispatch();
-  const userStore = useSelector((store) => store.userStore);
-  const { restaurantes } = useSelector((store) => store.restaurantStore);
+const Home = () => {
   const navigate = useNavigate();
-  const logOut = () => {
-    disptach(actionUserLogOutAsync());
-  };
-
-  useEffect(() => {
-    if (isAuthentication) {
-      setUserFunctions(true);
-    } else {
-      setUserFunctions(false);
-    }
-  }, [isAuthentication, userFunctions]);
-
-  // useEffect(() => {
-  // console.log(userStore);
-  // if (!userStore.name) {
-
-  //   navigate('/login')
-  //   console.log(userStore.name);
-  // }
-  // }, [userStore])
-
-  useEffect(() => {
-    if (!restaurantes.length) {
-      disptach(actionGetrestaurantesAsync());
-      console.log(restaurantes);
-    }
-  }, [restaurantes]);
-
-  console.log(restaurantes);
-  useEffect(() => {
-    console.log(userStore);
-  }, [userStore]);
-
-  const changeDado = () => {
-    setDado(true);
-    setTimeout(() => {
-      const x = Math.floor(Math.random() * lista.length);
-      console.log(x);
-      const y = lista.at(x);
-      console.log(y);
-      Swal.fire("tu restaurante sera", `${y}`, "info");
-    }, 1000);
-
-    setTimeout(() => {
-      setDado(false);
-    }, 2000);
-  };
   const handleDicePage = () => {
     navigate("/dice");
   };
-
   return (
     <>
       <NavbarDice />
@@ -108,25 +37,6 @@ const Home = ({ isAuthentication }) => {
           </button>
         </div>
       </section>
-
-      {isAuthentication ? <button onClick={logOut}> salir</button> : ""}
-      {isAuthentication && userStore.admin ? (
-        <div>
-          <button> Add restaurant</button>
-          <button> Edit restaurant</button>{" "}
-        </div>
-      ) : (
-        ""
-      )}
-      {isAuthentication && !userStore.admin ? (
-        <div>
-          <button> Favoritos</button>
-        </div>
-      ) : (
-        ""
-      )}
-      {isAuthentication ? userStore.displayName : ""}
-      {isAuthentication ? "" : <Link to={"/login"}>Ingresar </Link>}
       <hr />
       <div>
         <ControlledCarousel />
