+++ src/components/home/home.scss
@@ -44,6 +44,7 @@ $color-sesion: #f9943b;
     @include buttons(1);
     background-color: $color-sesion;
     margin-top: 100px;
+    margin-bottom: 100px;
   }
 }
 
@@ -157,7 +158,7 @@ hr {
   border: none;
   border-top: 3px solid black;
   margin: auto;
-  padding: 12px;
+  padding: 15px;
 }
 
 @media screen and (max-width: 500px) {
